# public/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "ProjectX Gateway API",
    "description": "ProjectX Gateway API Documentation and specification.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "produces": [
    "text/plain",
    "application/json",
    "text/json"
  ],
  "paths": {
...
    "/api/Order/search": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Searches for orders based on the provided request.",
        "operationId": "Order_SearchOrders",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing search criteria.",
            "schema": {
              "$ref": "#/definitions/SearchOrderRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response with search results.",
            "schema": {
              "$ref": "#/definitions/SearchOrderResponse"
            }
          }
        }
      }
    },
    "/api/Order/searchOpen": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Searches for open (working/active) orders based on the provided request.",
        "operationId": "Order_SearchOpenOrders",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing search criteria.",
            "schema": {
              "$ref": "#/definitions/SearchOpenOrderRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response with search results.",
            "schema": {
              "$ref": "#/definitions/SearchOrderResponse"
            }
          }
        }
      }
    },
    "/api/Order/place": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Places a new order based on the provided request.",
        "operationId": "Order_PlaceOrder",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing order details.",
            "schema": {
              "$ref": "#/definitions/PlaceOrderRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response with order placement details.",
            "schema": {
              "$ref": "#/definitions/PlaceOrderResponse"
            }
          }
        }
      }
    },
    "/api/Order/cancel": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Cancels an existing order based on the provided request.",
        "operationId": "Order_CancelOrder",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing order cancellation details.",
            "schema": {
              "$ref": "#/definitions/CancelOrderRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response with order cancellation details.",
            "schema": {
              "$ref": "#/definitions/CancelOrderResponse"
            }
          }
        }
      }
    },
    "/api/Order/modify": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Modifies an existing order based on the provided request.",
        "operationId": "Order_ModifyOrder",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing order modification details.",
            "schema": {
              "$ref": "#/definitions/ModifyOrderRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response with order modification details.",
            "schema": {
              "$ref": "#/definitions/ModifyOrderResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SearchOrderResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/SearchOrderErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderModel"
          }
        }
      }
    },
    "SearchOrderErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound",
      "x-enumNames": [
        "Success",
        "AccountNotFound"
      ],
      "enum": [
        0,
        1
      ]
    },
    "OrderModel": {
      "type": "object",
      "required": [
        "id",
        "accountId",
        "contractId",
        "symbolId",
        "creationTimestamp",
        "status",
        "type",
        "side",
        "size",
        "fillVolume"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "contractId": {
          "type": "string"
        },
        "symbolId": {
          "type": "string"
        },
        "creationTimestamp": {
          "type": "string",
          "format": "date-time"
        },
        "updateTimestamp": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "type": {
          "$ref": "#/definitions/OrderType"
        },
        "side": {
          "$ref": "#/definitions/OrderSide"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "limitPrice": {
          "type": "number",
          "format": "decimal"
        },
        "stopPrice": {
          "type": "number",
          "format": "decimal"
        },
        "fillVolume": {
          "type": "integer",
          "format": "int32"
        },
        "filledPrice": {
          "type": "number",
          "format": "decimal"
        },
        "customTag": {
          "type": "string"
        }
      }
    },
    "OrderStatus": {
      "type": "integer",
      "description": "0 = None\n1 = Open\n2 = Filled\n3 = Cancelled\n4 = Expired\n5 = Rejected\n6 = Pending\n7 = PendingCancellation\n8 = Suspended",
      "x-enumNames": [
        "None",
        "Open",
        "Filled",
        "Cancelled",
        "Expired",
        "Rejected",
        "Pending",
        "PendingCancellation",
        "Suspended"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8
      ]
    },
    "OrderType": {
      "type": "integer",
      "description": "0 = Unknown\n1 = Limit\n2 = Market\n3 = StopLimit\n4 = Stop\n5 = TrailingStop\n6 = JoinBid\n7 = JoinAsk",
      "x-enumNames": [
        "Unknown",
        "Limit",
        "Market",
        "StopLimit",
        "Stop",
        "TrailingStop",
        "JoinBid",
        "JoinAsk"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7
      ]
    },
    "OrderSide": {
      "type": "integer",
      "description": "0 = Bid\n1 = Ask",
      "x-enumNames": [
        "Bid",
        "Ask"
      ],
      "enum": [
        0,
        1
      ]
    },
    "SearchOrderRequest": {
      "type": "object",
      "required": [
        "accountId",
        "startTimestamp"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "startTimestamp": {
          "type": "string",
          "format": "date-time"
        },
        "endTimestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "SearchOpenOrderRequest": {
      "type": "object",
      "required": [
        "accountId"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "PlaceOrderResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/PlaceOrderErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "orderId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "PlaceOrderErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound\n2 = OrderRejected\n3 = InsufficientFunds\n4 = AccountViolation\n5 = OutsideTradingHours\n6 = OrderPending\n7 = UnknownError\n8 = ContractNotFound\n9 = ContractNotActive\n10 = AccountRejected",
      "x-enumNames": [
        "Success",
        "AccountNotFound",
        "OrderRejected",
        "InsufficientFunds",
        "AccountViolation",
        "OutsideTradingHours",
        "OrderPending",
        "UnknownError",
        "ContractNotFound",
        "ContractNotActive",
        "AccountRejected"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8,
        9,
        10
      ]
    },
    "PlaceOrderRequest": {
      "type": "object",
      "required": [
        "accountId",
        "contractId",
        "type",
        "side",
        "size"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "contractId": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/OrderType"
        },
        "side": {
          "$ref": "#/definitions/OrderSide"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "limitPrice": {
          "type": "number",
          "format": "decimal"
        },
        "stopPrice": {
          "type": "number",
          "format": "decimal"
        },
        "trailPrice": {
          "type": "number",
          "format": "decimal"
        },
        "customTag": {
          "type": "string"
        },
        "stopLossBracket": {
          "$ref": "#/definitions/PlaceOrderBracket"
        },
        "takeProfitBracket": {
          "$ref": "#/definitions/PlaceOrderBracket"
        }
      }
    },
    "PlaceOrderBracket": {
      "type": "object",
      "required": [
        "ticks",
        "type"
      ],
      "properties": {
        "ticks": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "$ref": "#/definitions/OrderType"
        }
      }
    },
    "CancelOrderResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/CancelOrderErrorCode"
        },
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "CancelOrderErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound\n2 = OrderNotFound\n3 = Rejected\n4 = Pending\n5 = UnknownError\n6 = AccountRejected",
      "x-enumNames": [
        "Success",
        "AccountNotFound",
        "OrderNotFound",
        "Rejected",
        "Pending",
        "UnknownError",
        "AccountRejected"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ]
    },
    "CancelOrderRequest": {
      "type": "object",
      "required": [
        "accountId",
        "orderId"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "orderId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "ModifyOrderResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/ModifyOrderErrorCode"
        },
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "ModifyOrderErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound\n2 = OrderNotFound\n3 = Rejected\n4 = Pending\n5 = UnknownError\n6 = AccountRejected\n7 = ContractNotFound",
      "x-enumNames": [
        "Success",
        "AccountNotFound",
        "OrderNotFound",
        "Rejected",
        "Pending",
        "UnknownError",
        "AccountRejected",
        "ContractNotFound"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7
      ]
    },
    "ModifyOrderRequest": {
      "type": "object",
      "required": [
        "accountId",
        "orderId"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "orderId": {
          "type": "integer",
          "format": "int64"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "limitPrice": {
          "type": "number",
          "format": "decimal"
        },
        "stopPrice": {
          "type": "number",
          "format": "decimal"
        },
        "trailPrice": {
          "type": "number",
          "format": "decimal"
        }
      }
    },
...
  }
}